Add test for happypath
